# docs/swagger.yaml
definitions:
  note.Note:
    properties:
      color:
        type: string
      laziness:
        type: integer
      name:
        type: string
      owner:
        type: string
      weight:
        type: number
    type: object
info:
  contact: {}
paths:
  /healthz:
    get:
      summary: Returns health of the service
  /note:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/note.Note'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get note list
      tags:
      - Note
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/note.Note'
      produces:
      - application/json
      responses:
        "200":
          description: successfully note created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new note
      tags:
      - Note
  /note/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: nil
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete note
      tags:
      - Note
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/note.Note'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get note
      tags:
      - Note
    put:
      consumes:
      - application/json
      description: Update note
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Update note
      tags:
      - Note
swagger: "2.0"